Replace config runner in commands
